Working on the Trie and Added Singleton pattern
I am currently working on the Trie and trying to make it properly
usable. I have to make it work for any type of char (from lowercase only
to UNICODE).
I implemented the Singleton pattern.
Once I am done with implementing this I will move to sorting algorithms
and binary_Search and then create a
DatabaseController class which will handle accessing to a MySQL Server.